Simplified code hooking mechanism
The core handler now steps the original hooked instruction before
calling the custom hook handler. This avoids the need to perform
the missing functionality in each handler. The core handler also
has no cycle cost, since the original instruction is run.

In most cases the hook handlers are shorter. In a few places the
hook needed to be adjusted for the new behaviour, but they perform
the same function as before.

Common behaviour was factored out to a shared lambda in DD + DD2.